# api/underfish-modified-api.openapi.yaml
openapi: 3.0.3
info:
    title: Underfish Platform API
    description: |
        Полная OpenAPI 3.0 спецификация публичных и внутренних HTTP ручек Underfish Platform,
        собранная по контроллерам authService, profileService, communityService, eventService,
        locationService, fileStorageService, gateway routes и notification endpoints из legacy underfish.

        Основной публичный вход — gateway `http://localhost:8080`. Профили публично доступны как
        `/api/v1/users/{user_id}`; gateway переписывает этот путь в profileService на
        `/api/v1/profiles/{user_id}`. Внутренние сервисные URL оставлены в `servers` для локальной
        отладки прямых сервисных вызовов.
    version: 2.0.0
servers:
    - url: http://localhost:8080
      description: Local API Gateway
    - url: https://underfish.ru
      description: Production API Gateway
    - url: http://localhost:8091
      description: Auth Service direct
    - url: http://localhost:8082
      description: Profile Service direct
    - url: http://localhost:8084
      description: Community Service direct
    - url: http://localhost:8083
      description: Event Service direct
    - url: http://localhost:8081
      description: Location Service direct
    - url: http://localhost:8085
      description: File Storage Service direct
...
paths:
...
    /api/v1/communities:
        post:
            tags: [Communities]
            summary: Создать сообщество
            operationId: createCommunity
            security: [{ bearerAuth: [] }]
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CommunityRequest'
            responses:
                '201':
                    description: Сообщество создано
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CommunityResponse'
                '401': { $ref: '#/components/responses/Unauthorized' }
                '403': { $ref: '#/components/responses/Forbidden' }
        get:
            tags: [Communities]
            summary: Получить список сообществ
            operationId: listCommunities
            parameters:
                - name: name
                  in: query
                  schema: { type: string }
                  description: Поиск по названию.
                - name: is_private
                  in: query
                  schema: { type: boolean }
                - name: tag_ids
                  in: query
                  description: Принимается для совместимости API, текущая версия communityService не использует фильтр.
                  style: form
                  explode: true
                  schema:
                      type: array
                      items: { type: string, format: uuid }
                - $ref: '#/components/parameters/Page'
                - $ref: '#/components/parameters/Size20'
                - name: sort
                  in: query
                  schema:
                      type: string
                      default: created_at
                      enum: [created_at, updated_at, name, status]
                - name: sort_direction
                  in: query
                  schema:
                      type: string
                      default: desc
                      enum: [asc, desc]
            responses:
                '200':
                    description: Страница сообществ
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CommunityPageResponse'
    /api/v1/communities/{community_id}:
        parameters:
            - $ref: '#/components/parameters/CommunityId'
        get:
            tags: [Communities]
            summary: Получить сообщество по ID
            operationId: getCommunity
            responses:
                '200':
                    description: Сообщество
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CommunityResponse'
                '404': { $ref: '#/components/responses/NotFound' }
        put:
            tags: [Communities]
            summary: Обновить сообщество
            operationId: updateCommunity
            security: [{ bearerAuth: [] }]
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CommunityRequest'
            responses:
                '200':
                    description: Сообщество обновлено
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CommunityResponse'
                '403': { $ref: '#/components/responses/Forbidden' }
                '404': { $ref: '#/components/responses/NotFound' }
        delete:
            tags: [Communities]
            summary: Удалить сообщество
            operationId: deleteCommunity
            security: [{ bearerAuth: [] }]
            responses:
                '204': { description: Сообщество удалено }
                '403': { $ref: '#/components/responses/Forbidden' }
                '404': { $ref: '#/components/responses/NotFound' }
...
    /api/v1/events:
        post:
            tags: [Events]
            summary: Создать мероприятие
            operationId: createEvent
            security: [{ bearerAuth: [] }]
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/EventRequest'
                        examples:
                            default:
                                value:
                                    title: E2E Event
                                    description: Event created by integration test
                                    start_datetime: '2030-01-01T10:00:00'
                                    end_datetime: '2030-01-01T12:00:00'
                                    location_id: '1'
                                    price: 0
                                    currency: RUB
                                    community_id: 72bb6a64-fd1b-4a7e-9649-80f91b261596
                                    status: PUBLISHED
                                    max_participants: 10
                                    is_online: true
            responses:
                '201':
                    description: Мероприятие создано
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/EventResponse'
                '400': { $ref: '#/components/responses/BadRequest' }
                '401': { $ref: '#/components/responses/Unauthorized' }
        get:
            tags: [Events]
            summary: Найти мероприятия
            operationId: listEvents
            parameters:
                - name: title
                  in: query
                  schema: { type: string }
                - name: status
                  in: query
                  schema:
                      type: string
                      enum: [DRAFT, PUBLISHED, CANCELLED, COMPLETED]
                - name: is_online
                  in: query
                  schema: { type: boolean }
                - name: start_from
                  in: query
                  schema: { type: string, format: date-time }
                - name: start_to
                  in: query
                  schema: { type: string, format: date-time }
                - name: price_min
                  in: query
                  schema: { type: number, format: double }
                - name: price_max
                  in: query
                  schema: { type: number, format: double }
                - name: location_id
                  in: query
                  description: Внешний ID локации. Сейчас хранится строкой и может быть числовым значением вида `1`.
                  schema: { type: string, example: '1' }
                - name: tag_ids
                  in: query
                  style: form
                  explode: true
                  schema:
                      type: array
                      items: { type: string, format: uuid }
                - name: community_id
                  in: query
                  schema: { type: string, format: uuid }
                - name: organizer_id
                  in: query
                  schema: { type: string, format: uuid }
                - $ref: '#/components/parameters/Page'
                - $ref: '#/components/parameters/Size20'
                - name: sort
                  in: query
                  schema:
                      type: string
                      default: startDatetime
                - name: sort_direction
                  in: query
                  schema:
                      type: string
                      default: asc
                      enum: [asc, desc]
            responses:
                '200':
                    description: Страница мероприятий
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/EventPageResponse'
    /api/v1/events/{event_id}:
        parameters:
            - $ref: '#/components/parameters/EventId'
        get:
            tags: [Events]
            summary: Получить мероприятие по ID
            operationId: getEvent
            responses:
                '200':
                    description: Мероприятие
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/EventResponse'
                '404': { $ref: '#/components/responses/NotFound' }
        put:
            tags: [Events]
            summary: Обновить мероприятие
            operationId: updateEvent
            security: [{ bearerAuth: [] }]
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/EventRequest'
            responses:
                '200':
                    description: Мероприятие обновлено
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/EventResponse'
                '403': { $ref: '#/components/responses/Forbidden' }
                '404': { $ref: '#/components/responses/NotFound' }
        delete:
            tags: [Events]
            summary: Удалить мероприятие
            operationId: deleteEvent
            security: [{ bearerAuth: [] }]
            responses:
                '204': { description: Мероприятие удалено }
                '403': { $ref: '#/components/responses/Forbidden' }
                '404': { $ref: '#/components/responses/NotFound' }
    /api/v1/events/{event_id}/tags:
        parameters:
            - $ref: '#/components/parameters/EventId'
        post:
            tags: [Events]
            summary: Привязать тег к мероприятию
            operationId: addTagToEvent
            security: [{ bearerAuth: [] }]
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/EventTagRequest'
            responses:
                '201':
                    description: Тег привязан
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/EventTagResponse'
        get:
            tags: [Events]
            summary: Получить теги мероприятия
            operationId: listEventTags
            responses:
                '200':
                    description: Список тегов мероприятия
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/EventTagResponse'
    /api/v1/events/{event_id}/tags/{tag_id}:
        delete:
            tags: [Events]
            summary: Отвязать тег от мероприятия
            operationId: removeTagFromEvent
            security: [{ bearerAuth: [] }]
            parameters:
                - $ref: '#/components/parameters/EventId'
                - $ref: '#/components/parameters/TagId'
            responses:
                '204': { description: Тег отвязан }
                '404': { $ref: '#/components/responses/NotFound' }
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    parameters:
...
        CommunityId:
            name: community_id
            in: path
            required: true
            schema: { type: string, format: uuid }
        EventId:
            name: event_id
            in: path
            required: true
            schema: { type: string, format: uuid }
        TagId:
            name: tag_id
            in: path
            required: true
            schema: { type: string, format: uuid }
...
        Page:
            name: page
            in: query
            schema: { type: integer, minimum: 0, default: 0 }
        Size20:
            name: size
            in: query
            schema: { type: integer, minimum: 1, default: 20 }
...
    responses:
        BadRequest:
            description: Некорректный запрос
            content:
                application/json:
                    schema: { $ref: '#/components/schemas/ErrorResponse' }
        Unauthorized:
            description: Требуется авторизация или токен невалиден
            content:
                application/json:
                    schema: { $ref: '#/components/schemas/ErrorResponse' }
        Forbidden:
            description: Недостаточно прав
            content:
                application/json:
                    schema: { $ref: '#/components/schemas/ErrorResponse' }
        NotFound:
            description: Ресурс не найден
            content:
                application/json:
                    schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    schemas:
...
        CommunityRequest:
            type: object
            required: [name, description]
            description: communityService сейчас не задаёт SNAKE_CASE, поэтому body fields wire-format в camelCase.
            properties:
                name: { type: string }
                description: { type: string }
                isPrivate: { type: boolean, default: false }
                coverUrl: { type: string, nullable: true }
        CommunityResponse:
            type: object
            required: [communityId, name, organizerId, isPrivate, createdAt, updatedAt, status]
            properties:
                communityId: { type: string, format: uuid }
                name: { type: string }
                description: { type: string, nullable: true }
                organizerId: { type: string, format: uuid }
                isPrivate: { type: boolean }
                coverUrl: { type: string, nullable: true }
                createdAt: { type: string, format: date-time }
                updatedAt: { type: string, format: date-time }
                status: { type: string, enum: [ACTIVE, SUSPENDED, ARCHIVED] }
        CommunityPageResponse:
            type: object
            required: [content, meta]
            properties:
                content:
                    type: array
                    items: { $ref: '#/components/schemas/CommunityResponse' }
                meta: { $ref: '#/components/schemas/CommunityPageMeta' }
        CommunityPageMeta:
            type: object
            required: [page, size, totalElements, totalPages]
            properties:
                page: { type: integer }
                size: { type: integer }
                totalElements: { type: integer }
                totalPages: { type: integer }
...
        EventRequest:
            type: object
            required: [title, description, start_datetime, end_datetime, location_id, status]
            properties:
                title: { type: string, minLength: 1 }
                description: { type: string, minLength: 1 }
                start_datetime: { type: string, format: date-time }
                end_datetime: { type: string, format: date-time }
                location_id:
                    type: string
                    description: Внешний ID locationService; хранится как текст и может быть числовой строкой.
                    example: '1'
                price: { type: number, format: double, nullable: true }
                currency: { type: string, nullable: true }
                poster_url: { type: string, nullable: true }
                community_id: { type: string, format: uuid, nullable: true }
                status: { type: string, enum: [DRAFT, PUBLISHED, CANCELLED, COMPLETED] }
                max_participants: { type: integer, nullable: true }
                is_online: { type: boolean, nullable: true }
        EventResponse:
            type: object
            required: [event_id, organizer_id, title, description, start_datetime, location_id, price, currency, status, created_at, updated_at, max_participants, is_online]
            properties:
                event_id: { type: string, format: uuid }
                organizer_id: { type: string, format: uuid }
                title: { type: string }
                description: { type: string }
                start_datetime: { type: string, format: date-time }
                end_datetime: { type: string, format: date-time, nullable: true }
                location_id: { type: string, example: '1' }
                price: { type: number, format: double }
                currency: { type: string }
                poster_url: { type: string, nullable: true }
                community_id: { type: string, format: uuid, nullable: true }
                status: { type: string, enum: [DRAFT, PUBLISHED, CANCELLED, COMPLETED] }
                created_at: { type: string, format: date-time }
                updated_at: { type: string, format: date-time }
                max_participants: { type: integer }
                is_online: { type: boolean }
        EventPageResponse:
            type: object
            required: [content, meta]
            properties:
                content:
                    type: array
                    items: { $ref: '#/components/schemas/EventResponse' }
                meta: { $ref: '#/components/schemas/PageMeta' }
        EventTagRequest:
            type: object
            required: [tag_id]
            properties:
                tag_id: { type: string, format: uuid }
        EventTagResponse:
            type: object
            required: [event_id, tag_id]
            properties:
                event_id: { type: string, format: uuid }
                tag_id: { type: string, format: uuid }
...
        PageMeta:
            type: object
            required: [page, size, total_elements, total_pages]
            properties:
                page: { type: integer }
                size: { type: integer }
                total_elements: { type: integer, format: int64 }
                total_pages: { type: integer }
        ErrorResponse:
            type: object
            properties:
                timestamp: { type: string, format: date-time }
                status: { type: integer }
                error: { type: string }
                message: { type: string, nullable: true }
                path: { type: string, nullable: true }