# profileService/src/main/kotlin/ru/underfish/profile_service/dto/response/ErrorResponse.kt
package ru.underfish.profile_service.dto.response

import java.time.LocalDateTime

data class ErrorResponse(
    val timestamp: LocalDateTime = LocalDateTime.now(),
    val status: Int,
    val error: String,
    val message: String?,
    val path: String?,
)